+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/messenger/compose/view/room/settings/RoomSettingsLeaveRoom.kt
@@ -12,14 +12,15 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.text.capitalize
-import androidx.compose.ui.text.intl.Locale
 import de.connect2x.messenger.compose.view.DI
-import de.connect2x.messenger.compose.view.common.WarningDialog
 import de.connect2x.messenger.compose.view.get
 import de.connect2x.messenger.compose.view.i18n.I18nView
 import de.connect2x.messenger.compose.view.theme.components
+import de.connect2x.messenger.compose.view.theme.components.ModalDialogContent
+import de.connect2x.messenger.compose.view.theme.components.ModalDialogFooter
+import de.connect2x.messenger.compose.view.theme.components.ModalDialogHeader
 import de.connect2x.messenger.compose.view.theme.components.ThemedButton
+import de.connect2x.messenger.compose.view.theme.components.ThemedModalDialog
 import de.connect2x.trixnity.messenger.viewmodel.room.settings.RoomSettingsViewModel
 
 interface RoomSettingsLeaveRoomView {
@@ -55,25 +56,40 @@ class RoomSettingsLeaveRoomViewImpl : RoomSettingsLeaveRoomView {
 @Composable
 fun RoomSettingsLeaveRoomWarning(roomSettingsViewModel: RoomSettingsViewModel) {
     val i18n = DI.get<I18nView>()
-    val isLeave = roomSettingsViewModel.isLeave.collectAsState().value
-    val leaveRoomWarningMessage = roomSettingsViewModel.leaveRoomWarningMessage.collectAsState().value
-    val leaveRoomWarningTitle = roomSettingsViewModel.leaveRoomWarningTitle.collectAsState().value
-    val leaveRoomWarningConfirmButtonText =
-        roomSettingsViewModel.leaveRoomWarningConfirmButtonText.collectAsState().value
 
-    WarningDialog(
-        title = leaveRoomWarningTitle,
-        message = { Text(leaveRoomWarningMessage) },
-        dismissButtonText = i18n.commonCancel().capitalize(Locale.current),
-        confirmButtonText = leaveRoomWarningConfirmButtonText,
-        dismissAction = { roomSettingsViewModel.closeLeaveRoomWarningDialog() },
-        confirmAction = {
-            if (isLeave) {
-                roomSettingsViewModel.forgetRoom()
-            } else {
-                roomSettingsViewModel.leaveRoom()
-            }
-            roomSettingsViewModel.closeLeaveRoomWarningDialog()
+    val isLeave = roomSettingsViewModel.isLeave.collectAsState().value
+
+    val title = roomSettingsViewModel.leaveRoomWarningTitle.collectAsState().value
+    val message = roomSettingsViewModel.leaveRoomWarningMessage.collectAsState().value
+    val confirm = roomSettingsViewModel.leaveRoomWarningConfirmButtonText.collectAsState().value
+
+    ThemedModalDialog({ roomSettingsViewModel.closeLeaveRoomWarningDialog() }) {
+        ModalDialogHeader {
+            Text(title)
         }
-    )
+        ModalDialogContent {
+            Text(message)
+        }
+        ModalDialogFooter {
+            ThemedButton(
+                style = MaterialTheme.components.commonButton,
+                onClick = { roomSettingsViewModel.closeLeaveRoomWarningDialog() },
+            ) {
+                Text(i18n.actionCancel())
+            }
+            ThemedButton(
+                style = MaterialTheme.components.primaryButton,
+                onClick = {
+                    if (isLeave) {
+                        roomSettingsViewModel.forgetRoom()
+                    } else {
+                        roomSettingsViewModel.leaveRoom()
+                    }
+                    roomSettingsViewModel.closeLeaveRoomWarningDialog()
+                }
+            ) {
+                Text(confirm)
+            }
+        }
+    }
 }