+++ trixnity-messenger/src/androidMain/kotlin/de/connect2x/trixnity/messenger/util/Images.android.kt
@@ -3,8 +3,19 @@ package de.connect2x.trixnity.messenger.util
 import android.graphics.BitmapFactory
 import de.connect2x.trixnity.messenger.viewmodel.util.limitedByteArrayOrNull
 import net.folivo.trixnity.utils.ByteArrayFlow
+import org.koin.core.module.Module
+import org.koin.dsl.module
 
-actual suspend fun getImageDimensions(byteArrayFlow: ByteArrayFlow, maxMediaSize: Long): Pair<Int?, Int?> {
+actual fun platformGetImageDimensionsModule(): Module = module {
+    single<GetImageDimensions> {
+        GetImageDimensions { byteArrayFlow, maxSize ->
+            getImageDimensions(byteArrayFlow, maxSize)
+        }
+    }
+}
+
+
+suspend fun getImageDimensions(byteArrayFlow: ByteArrayFlow, maxMediaSize: Long): Pair<Int?, Int?> {
     val byteArray = byteArrayFlow.limitedByteArrayOrNull(maxMediaSize)
     val bitmap = byteArray?.let { BitmapFactory.decodeByteArray(it, 0, it.size) }
     return bitmap?.width to bitmap?.height