+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/message/EncryptedErrorTimelineElementViewModel.kt
@@ -1,10 +1,15 @@
-package de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements
+package de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.message
 
 import de.connect2x.trixnity.messenger.i18n.I18n
 import de.connect2x.trixnity.messenger.viewmodel.MatrixClientViewModelContext
+import de.connect2x.trixnity.messenger.viewmodel.room.timeline.elements.TimelineElementViewModel.Message
 import net.folivo.trixnity.client.store.TimelineEvent.TimelineEventContentError
 import net.folivo.trixnity.core.model.events.m.room.EncryptedMessageEventContent
 import org.koin.core.component.get
+
+interface EncryptedErrorTimelineElementViewModel : Message<EncryptedMessageEventContent> {
+    val error: String
+}
 
 interface EncryptedErrorTimelineElementViewModelFactory {
     fun create(
@@ -20,14 +25,13 @@ interface EncryptedErrorTimelineElementViewModelFactory {
     companion object : EncryptedErrorTimelineElementViewModelFactory
 }
 
-interface EncryptedErrorTimelineElementViewModel : TimelineElementViewModel.Message<EncryptedMessageEventContent>
 
 class EncryptedErrorTimelineElementViewModelImpl(
     viewModelContext: MatrixClientViewModelContext,
     error: Throwable,
 ) : EncryptedErrorTimelineElementViewModel, MatrixClientViewModelContext by viewModelContext {
     private val i18n = get<I18n>()
-    val error = when (error) {
+    override val error = when (error) {
         TimelineEventContentError.DecryptionAlgorithmNotSupported -> i18n.timelineElementDecryptionErrorAlgorithmNotSupported()
         is TimelineEventContentError.DecryptionError -> i18n.timelineElementDecryptionErrorGeneric(error.error.message)
         TimelineEventContentError.DecryptionTimeout -> i18n.timelineElementDecryptionErrorTimeout()