+++ trixnity-messenger/src/commonTest/kotlin/de/connect2x/trixnity/messenger/viewmodel/room/timeline/elements/state/MemberStateTimelineElementViewModelTest.kt
@@ -161,6 +161,7 @@ class MemberStateTimelineElementViewModelTest {
     fun `leaving user » should show an indicator for user leaving a room`() = runTest {
         val cut = memberStatusViewModel(
             mockTimelineEvent(
+                senderId = UserId("@bob:localhost"),
                 membership = Membership.LEAVE,
                 previousMemberEventContent = memberEventContent(membership = Membership.JOIN),
                 stateKey = "@bob:localhost",
@@ -183,7 +184,7 @@ class MemberStateTimelineElementViewModelTest {
         )
         backgroundScope.launch { cut.changeMessage.collect {} }
         eventually(100.milliseconds) {
-            cut.changeMessage.value shouldBe "Mallory has been removed by Sender from the group"
+            cut.changeMessage.value shouldBe "Mallory was banned from the group by Sender"
         }
     }
 
@@ -199,7 +200,69 @@ class MemberStateTimelineElementViewModelTest {
         )
         backgroundScope.launch { cut.changeMessage.collect {} }
         eventually(100.milliseconds) {
-            cut.changeMessage.value shouldBe "Mallory has been removed by Sender from the group because \"he spammed our chat :(\""
+            cut.changeMessage.value shouldBe "Mallory was banned from the group by Sender because \"he spammed our chat :(\""
+        }
+    }
+
+    @Test
+    fun `banned user » should show an indicator for user being unbanned from a room`() = runTest {
+        val cut = memberStatusViewModel(
+            mockTimelineEvent(
+                membership = Membership.LEAVE,
+                previousMemberEventContent = memberEventContent(membership = Membership.BAN),
+                stateKey = "@mallory:localhost",
+            ),
+        )
+        backgroundScope.launch { cut.changeMessage.collect {} }
+        eventually(100.milliseconds) {
+            cut.changeMessage.value shouldBe "Mallory was unbanned by Sender"
+        }
+    }
+
+    @Test
+    fun `banned user » should show an indicator with reason for user being unbanned from a room`() = runTest {
+        val cut = memberStatusViewModel(
+            mockTimelineEvent(
+                membership = Membership.LEAVE,
+                previousMemberEventContent = memberEventContent(membership = Membership.BAN),
+                stateKey = "@mallory:localhost",
+                reason = "he spammed our chat :("
+            ),
+        )
+        backgroundScope.launch { cut.changeMessage.collect {} }
+        eventually(100.milliseconds) {
+            cut.changeMessage.value shouldBe "Mallory was unbanned by Sender because \"he spammed our chat :(\""
+        }
+    }
+
+    @Test
+    fun `kicked user » should show an indicator for user being kicked from a room`() = runTest {
+        val cut = memberStatusViewModel(
+            mockTimelineEvent(
+                membership = Membership.LEAVE,
+                previousMemberEventContent = memberEventContent(membership = Membership.JOIN),
+                stateKey = "@mallory:localhost",
+            ),
+        )
+        backgroundScope.launch { cut.changeMessage.collect {} }
+        eventually(100.milliseconds) {
+            cut.changeMessage.value shouldBe "Mallory was removed from the group by Sender"
+        }
+    }
+
+    @Test
+    fun `kicked user » should show an indicator with reason for user being kicked from a room`() = runTest {
+        val cut = memberStatusViewModel(
+            mockTimelineEvent(
+                membership = Membership.LEAVE,
+                previousMemberEventContent = memberEventContent(membership = Membership.JOIN),
+                stateKey = "@mallory:localhost",
+                reason = "he spammed our chat :("
+            ),
+        )
+        backgroundScope.launch { cut.changeMessage.collect {} }
+        eventually(100.milliseconds) {
+            cut.changeMessage.value shouldBe "Mallory was removed from the group by Sender because \"he spammed our chat :(\""
         }
     }
 
@@ -214,7 +277,7 @@ class MemberStateTimelineElementViewModelTest {
         )
         backgroundScope.launch { cut.changeMessage.collect {} }
         eventually(100.milliseconds) {
-            cut.changeMessage.value shouldBe "Bob has been invited by Sender"
+            cut.changeMessage.value shouldBe "Bob was invited by Sender"
         }
     }
 
@@ -230,7 +293,7 @@ class MemberStateTimelineElementViewModelTest {
         )
         backgroundScope.launch { cut.changeMessage.collect {} }
         eventually(100.milliseconds) {
-            cut.changeMessage.value shouldBe "Bob has been invited by Sender because \"I want him to play Stardew Valley with us\""
+            cut.changeMessage.value shouldBe "Bob was invited by Sender because \"I want him to play Stardew Valley with us\""
         }
     }
 
@@ -309,12 +372,12 @@ class MemberStateTimelineElementViewModelTest {
         )
         backgroundScope.launch { cut.changeMessage.collect {} }
         eventually(100.milliseconds) {
-            cut.changeMessage.value shouldBe "Bob has been invited by Sender"
+            cut.changeMessage.value shouldBe "Bob was invited by Sender"
         }
 
         senderName.value = "Sender2"
         eventually(100.milliseconds) {
-            cut.changeMessage.value shouldBe "Bob has been invited by Sender2"
+            cut.changeMessage.value shouldBe "Bob was invited by Sender2"
         }
     }